Actions: Introduce A11y test (#106806)
