+++ package.json
@@ -224,6 +224,7 @@
     "node-notifier": "10.0.1",
     "nx": "20.7.1",
     "openapi-types": "^12.1.3",
+    "pa11y-ci": "^3.1.0",
     "pdf-parse": "^1.1.1",
     "plop": "^4.0.1",
     "postcss": "8.5.1",
