+++ scripts/grafana-server/kill-server
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 . scripts/grafana-server/variables
