+++ scripts/grafana-server/start-server
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 set -eo pipefail
 
 . scripts/grafana-server/variables
