+++ scripts/grafana-server/wait-for-grafana
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 set -eo pipefail
 
 . scripts/grafana-server/variables
